@@ -18,16 +18,16 @@ repositories {
 }
 
 dependencies {
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'com.android.support:design:27.+'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation 'com.github.KalebKE:FSensor:v1.0'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.1'
-    implementation "com.android.support:support-v4:27.+"
-    implementation "com.android.support:support-v13:27.+"
-    implementation "com.android.support:cardview-v7:27.+"
-    implementation "com.android.support:appcompat-v7:27.+"
-    implementation "android.arch.lifecycle:extensions:1.+"
-    annotationProcessor "android.arch.lifecycle:compiler:1.+"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.0.0'
 }
 
 // The sample build uses multiple directories to
@@ -39,18 +39,18 @@ List<String> dirs = [
         'template'] // boilerplate code that is generated by the sample template process
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 33
 
     buildToolsVersion "27.0.3"
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 33
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     sourceSets {
