@@ -1,7 +1,7 @@
 package com.example.android.camera2basic.ui;
 
 import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.example.android.camera2basic.R;
